requirements.txt
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
marshmallow-sqlalchemy==0.29.0
SQLAlchemy==2.0.8
psycopg2==2.9.5 # windows
# psycopg2-binary # linux
